Move SelectCoins max_weight test
