+++ src/Makefile.test.include
@@ -181,7 +181,6 @@ BITCOIN_TESTS += \
   wallet/test/walletdb_tests.cpp \
   wallet/test/wallet_crypto_tests.cpp \
   wallet/test/wallet_transaction_tests.cpp \
-  wallet/test/coinselector_tests.cpp \
   wallet/test/coinselection_tests.cpp \
   wallet/test/init_tests.cpp \
   wallet/test/ismine_tests.cpp \
